requirements.txt
behave
compare
pyyaml
requests
pylint